# services/trading-service/docs/swagger.yaml
definitions:
...
  errors.AppError:
    properties:
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
info:
  contact: {}
  description: API for managing portfolios, executing trades, and handling financial
    market operations.
  title: Trading Service API
  version: "1.0"
paths:
...
  /api/client/{clientId}/assets/{id}/publish/{amount}:
    patch:
      description: Sets the number of assets the caller makes publicly visible on
        the OTC portal.
      parameters:
      - description: Asset ownership ID
        in: path
        name: id
        required: true
        type: integer
      - description: Amount to make public
        in: path
        name: amount
        required: true
        type: number
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: Publish assets for OTC trading
      tags:
      - otc
...
  /api/otc/public:
    get:
      description: Returns a paginated list of assets that have been marked public
        by their owners on the OTC portal.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - BearerAuth: []
      summary: List all publicly available OTC assets
      tags:
      - otc
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"